shieldTV_loader: clean up usbd.* a bit
I want to keep these as generic as possible, without
intertwining it with the rest of shieldTV_loader.

Okay I am almost kind of happy with it - it makes no assumptions
about the environment at all. No allocations. If/when I decide
shieldTV_loader needs the complexity of enabling an MMU, the data cache
ops will have to be added to the usbd context.

One last real TODO: for every while (wait_until_hw_done) add
a break-out condition on USBSTS SLI/RESET.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>